[readability] move get_rnn_cell out of seq2seq_encoder_decoder (#1073)
* [readability] move get_rnn_cell out of seq2seq_encoder_decoder

* fix mport

* Update seq2seq_encoder_decoder.py

* Update utils.py
