+++ src/gluonnlp/model/attention_cell.py
@@ -504,3 +504,46 @@ class DotProductAttentionCell(AttentionCell):
             att_weights = att_weights * mask
         att_weights = self._dropout_layer(att_weights)
         return att_weights
+
+def _get_attention_cell(attention_cell, units=None,
+                        scaled=True, num_heads=None,
+                        use_bias=False, dropout=0.0):
+    """
+
+    Parameters
+    ----------
+    attention_cell : AttentionCell or str
+    units : int or None
+
+    Returns
+    -------
+    attention_cell : AttentionCell
+    """
+    if isinstance(attention_cell, str):
+        if attention_cell == 'scaled_luong':
+            return DotProductAttentionCell(units=units, scaled=True, normalized=False,
+                                           use_bias=use_bias, dropout=dropout, luong_style=True)
+        elif attention_cell == 'scaled_dot':
+            return DotProductAttentionCell(units=units, scaled=True, normalized=False,
+                                           use_bias=use_bias, dropout=dropout, luong_style=False)
+        elif attention_cell == 'dot':
+            return DotProductAttentionCell(units=units, scaled=False, normalized=False,
+                                           use_bias=use_bias, dropout=dropout, luong_style=False)
+        elif attention_cell == 'cosine':
+            return DotProductAttentionCell(units=units, scaled=False, use_bias=use_bias,
+                                           dropout=dropout, normalized=True)
+        elif attention_cell == 'mlp':
+            return MLPAttentionCell(units=units, normalized=False)
+        elif attention_cell == 'normed_mlp':
+            return MLPAttentionCell(units=units, normalized=True)
+        elif attention_cell == 'multi_head':
+            base_cell = DotProductAttentionCell(scaled=scaled, dropout=dropout)
+            return MultiHeadAttentionCell(base_cell=base_cell, query_units=units, use_bias=use_bias,
+                                          key_units=units, value_units=units, num_heads=num_heads)
+        else:
+            raise NotImplementedError
+    else:
+        assert isinstance(attention_cell, AttentionCell),\
+            'attention_cell must be either string or AttentionCell. Received attention_cell={}'\
+                .format(attention_cell)
+        return attention_cell
