+++ src/gluonnlp/model/transformer.py
@@ -34,10 +34,10 @@ from ..base import get_home_dir
 from ..utils.parallel import Parallelizable
 from .block import GELU
 from .seq2seq_encoder_decoder import (Seq2SeqDecoder, Seq2SeqEncoder,
-                                      Seq2SeqOneStepDecoder,
-                                      _get_attention_cell)
+                                      Seq2SeqOneStepDecoder)
 from .translation import NMTModel
 from .utils import _load_pretrained_params, _load_vocab
+from .attention_cell import _get_attention_cell
 
 def _position_encoding_init(max_length, dim):
     """Init the sinusoid position encoding table """
